fix(store): widen int before Duration multiply
On V 0.5.2 cbf4e85, the compiler this project pins, `int * time.Unit`
truncates the product to 32 bits before widening it to the i64
time.Duration it assigns to. Every unit parse_duration produced was
wrong: --last 30s came out at -64771072 ns, --last 30d at 121765888 ns
(both negative or absurdly small instead of tens of billions), and
--timeout above roughly 2147 ms went negative the same way. My own
local V checkout is a newer, unpinned commit that does not have this
bug, so it passed every make check I ran this session; the pinned
compiler the release workflow actually builds with does not, and
that is what surfaced this cutting v0.1.0 just now.

The fix is the same in both places: widen the left operand to i64
before the multiplication chain, so every step happens at i64 width.
store/history_test.v's own expected values had the identical bug in
their literal expressions and needed the same fix to assert the
correct numbers rather than agreeing with the truncated ones.

Verified against V built at the pinned commit, not the one this
session had been using: make check is clean and store/history.v's
tests, parse_duration specifically, now pass under it.

+++ cmd/cli.v
@@ -707,7 +707,10 @@ fn parse_args(args []string) !Options {
 				if ms < 1 {
 					return error('--timeout needs a positive number of milliseconds')
 				}
-				o = Options{ ...o, timeout: ms * time.millisecond }
+				// i64 first: `ms * time.millisecond` truncates to a 32-bit
+				// product before widening to the i64 Duration on V 0.5.2
+				// cbf4e85, silently going negative anywhere above ~2147 ms.
+				o = Options{ ...o, timeout: i64(ms) * time.millisecond }
 			}
 			'--cold-zone' {
 				o = Options{ ...o, cold_zone: value }
